+++ glo/trunk/src/Texture1D.cpp
@@ -26,7 +26,7 @@ const bool Texture1D::isBound() const
 	GLint currentBinding;
 	glGetIntegerv( GL_TEXTURE_BINDING_1D, &currentBinding );
 
-	return ( currentBinding == m_object );
+	return ( static_cast<GLuint>(currentBinding) == m_object );
 }
